my second commit 10.02.2023

+++ src/main/java/motorola/example/app/DeviceController.java
@@ -1,8 +1,10 @@
 package motorola.example.app;
 
 import com.google.gson.Gson;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
@@ -13,6 +15,7 @@ import java.util.Optional;
 
 @Controller
 @ResponseBody
+@Transactional
 public class DeviceController {
 
     private final DeviceRepository deviceRepository;
@@ -32,20 +35,19 @@ public class DeviceController {
         return gson.toJson(listOfDevices);
     }
 
-    @GetMapping("/getdevicesbyid")
-    public String getDevicesById(@RequestParam(name = "id") Long id){
-        Optional<Device> device = deviceRepository.findById(id);
-        return device.toString();
-    }
 
-    @GetMapping("/getdevicesbyenabled")
+
+    @GetMapping("/getDevicesByEnabled")
     public String getDevicesByEnabled(@RequestParam(name = "enabled")String enabled){
-        List<Device> devices = deviceRepository.findAllByEnabled(enabled);
-        Gson gson = new Gson();
-        return gson.toJson(devices);
+        if(enabled.equals("true") || enabled.equals("false")) {
+            List<Device> devices = deviceRepository.findAllByEnabled(enabled);
+            Gson gson = new Gson();
+            return gson.toJson(devices);
+        }
+        return "Use 'false' or 'true'";
     }
 
-    @GetMapping("/getdevicesbyip")
+    @GetMapping("/getDevicesByIp")
     public String getDevicesByIp(@RequestParam(name = "ip")String ip){
         Device device = deviceRepository.findByIpAddress(ip);
         return device.toString();
@@ -63,24 +65,33 @@ public class DeviceController {
         return "Devices saved";
     }
 
-    @GetMapping("/deletebyid")
-    public String deleteById(@RequestParam(name = "id")Long id){
-        deviceRepository.deleteById(id);
-        return "Device deleted id: "+id;
-    }
 
-    @GetMapping("/deletebyhostname")
+
+    @GetMapping("/deleteByHostname")
     public String deleteByHostname(@RequestParam(name = "hostname") String hostname){
+        Device device = deviceRepository.findByHostname(hostname);
+        if(device == null){
+            return "Device with hostname "+hostname+" does not exist";
+        }
         deviceRepository.deleteByHostname(hostname);
+
 
         return "Device deleted hostname: "+hostname;
     }
 
-    @GetMapping("/updateenabled")
-    public String updateEnabled(@RequestParam(name = "id")Long id, @RequestParam(name = "enabled") String enabled){
-        Optional<Device> device = deviceRepository.findById(id);
-        device.ifPresent(e -> e.enabled(enabled));
-        deviceRepository.save(device.get());
-        return "Device updated";
+    @PutMapping("/updateEnabled")
+    public String updateEnabled(@RequestParam(name = "hostname")String hostname, @RequestParam(name = "enabled") String enabled){
+        Optional<Device> device = Optional.ofNullable(deviceRepository.findByHostname(hostname));
+        if(device.toString().equals("Optional.empty")){
+            return "Device does not exist";
+        }
+
+        if(enabled.equals("true") || enabled.equals("false")){
+            device.ifPresent(e -> e.enabled(enabled));
+            device.ifPresent(e -> deviceRepository.save(device.get()));
+            return "Device "+hostname+" updated";
+        }
+
+        return "Use 'false' or 'true'";
     }
 }
